Add Choose class to handle Rental person and book choices
